requirements.txt
absl-py==0.9.0
APScheduler==3.6.3
argcomplete==1.10.0
astroid==2.3.3
attrs==19.3.0
autopep8==1.5
beautifulsoup4==4.8.0
boto==2.49.0
boto3==1.11.11
botocore==1.14.11
bpemb==0.3.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
cloudpickle==1.2.2
cycler==0.10.0
Cython==0.29.14
decorator==4.4.1
Deprecated==1.2.7
detectron2==0.1
docutils==0.15.2
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
filelock==3.0.12
flair==0.4.5
Flask==1.1.1
Flask-Cors==3.0.8
Flask-JWT-Extended==3.24.1
Flask-PyMongo==2.3.0
future==0.18.2
fuzzywuzzy==0.17.0
fvcore==0.1.dev200206
gensim==3.8.1
google-auth==1.11.0
google-auth-oauthlib==0.4.1
grpcio==1.26.0
hyperopt==0.2.3
idna==2.8
IMAPClient==2.1.0
importlib-metadata==1.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
joblib==0.14.1
kiwisolver==1.1.0
langdetect==1.0.7
lazy-object-proxy==1.4.3
lxml==4.5.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
more-itertools==8.2.0
mpld3==0.3
networkx==2.2
numpy==1.18.1
oauthlib==3.1.0
olefile==0.46
opencv-python==4.2.0.32
packaging==20.1
pdf2image==1.11.0
pdfminer.six==20200124
Pillow==7.0.0
pkg-resources==0.0.0
pluggy==0.13.1
portalocker==1.5.2
protobuf==3.11.3
py==1.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0
pycodestyle==2.5.0
pycryptodome==3.9.6
pydot==1.4.1
PyJWT==1.7.1
pylint==2.4.4
pymongo==3.10.1
pyparsing==2.4.6
pytest==5.3.5
python-dateutil==2.8.1
python-Levenshtein==0.12.0
python-pptx==0.6.18
pytz==2019.3
PyYAML==5.3
regex==2020.1.8
requests==2.22.0
requests-oauthlib==1.3.0
rsa==4.0
s3transfer==0.3.2
sacremoses==0.0.38
scikit-learn==0.22.1
scipy==1.4.1
segtok==1.5.7
sentencepiece==0.1.85
six==1.12.0
smart-open==1.9.0
sortedcontainers==2.1.0
soupsieve==1.9.5
SpeechRecognition==3.8.1
sqlitedict==1.6.0
tabulate==0.8.6
tensorboard==2.1.0
termcolor==1.1.0
tesserocr==2.5.0
tokenizers==0.0.11
torch==1.4.0
torchvision==0.5.0
tqdm==4.42.1
transformers==2.4.1
transforms==0.1
typed-ast==1.4.1
tzlocal==1.5.1
urllib3==1.24.3
wcwidth==0.1.8
Werkzeug==0.16.1
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.7
yacs==0.1.6
zipp==2.1.0
